refactor(scripts): change deploy scripts names
